pyproject.toml
[tool.poetry]
name = "workshop-b2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^30.8.1"
fastapi = "^0.115.3"
httpx = "^0.27.2"
sqlmodel = "^0.0.22"
typer = "^0.12.5"
uvicorn = "^0.32.0"
strawberry-graphql = "^0.247.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
